feat: in-memory mode
This commit makes the default export of `sorcery` a function. You must pass an array whose items must be a string or an object. String items must be the source code of a file with an inline source map. Object items must have a `content` string property, and an optional `map` object property.

The `sorcery` function returns a SourceMap object. Its `file` property must be set before calling `toJSON` if its associated file will be saved somewhere on disk.

You can pass an options object as the second argument, which takes the
same options as `Chain.prototype.apply`.

Other changes made in this commit:
- add `sourceRoot` property to SourceMap
- add `decode` method to Node
- use process.cwd() when `node.file` is null
- check for `node.map` in the getMap function