refactor `sorcery` and `loadSources`
The `sourceRoot` option no longer has a default value.

The `generatedFile` option is now used without changes.

The `sourceRoot` and `generatedFile` options cannot be absolute
paths. Otherwise, an error is thrown. This is to discourage
fragile/leaky sourcemaps.

The `sourceRoot` property is excluded from the returned sourcemap
when its `sources` array only contains null. (Can this even happen?)

For the `sourceRoot` string of an existing sourcemap, make it
absolute if we can. Otherwise, the source paths will be relative,
which is annoying for the caller.

For the `sources` array of an existing sourcemap, assume its paths
are always relative.

Avoid tracing a new sourcemap if only two nodes were given and
one of them is the original source.

Account for null in the `sources` array when generating the
`sourcesContent` array.

Account for null when loading the `sources` array of an
existing sourcemap.