stop forcing `node.file` to be absolute
The caller can easily resolve file paths themself if they want, but it may be easier to work with relative paths instead.

Side effect: when `readFile` is called with a sourcemap path, it will contain /../ in the middle if the sourceMappingURL starts with ../

We should probably make this an option.