fix: interweaving between current and next
When deciding if an old segment can be interwoven, we need to compare the *generated* column of the next root segment with the *generated* and *shifted* column of the old segment.

Previously, we were comparing the *source* column of the next root segment with the *generated* column of the old segment. This caused old segments to be interwoven even though they didn't actually fit (eg: their generated column was larger than the generated column of the next segment in the final mappings).